+++ app/Resources/css/contact/contact.css
@@ -1,5 +1,7 @@
 @import '/assets/styles/app.css';
 @import '../home/header.css';
+@import './faq.section.css';
+@import './form.section.css';
 
 body {
   background-color: hsl(var(--background) / 0.6);
@@ -10,25 +12,6 @@ body {
   border-color: hsl(var(--border-theme) / 0.3);
 }
 
-
-.bg-contact {
-  background: linear-gradient(
-    135deg,
-    hsl(176, 43%, 45%) 0%,
-    hsl(176, 43%, 47%) 40%,
-    hsl(176, 43%, 53%) 60%,
-    hsl(176, 43%, 55%) 100%
-  );
-  color: #F5F5F5;
-  border-bottom: 8px solid transparent;
-  border-image: repeating-linear-gradient(
-    to right,
-    #fafafa 0 10px,
-    transparent 10px 20px
-  );
-  border-image-slice: 1;
-}
-
 .banner{
   background: linear-gradient(
     135deg,
@@ -37,32 +20,6 @@ body {
     hsl(176, 43%, 53%) 60%,
     hsl(176, 43%, 55%) 100%
   );
-}
-
-@media (min-width: 768px) {
-  .bg-contact{
-      border-bottom: 0px;
-      border-right: 8px solid transparent;
-      border-image: repeating-linear-gradient(
-        to bottom,
-        #fafafa 0 10px,
-        transparent 10px 20px
-      );
-      border-image-slice: 1;
-  }
-}
-
-.faq-text{
-  height: 0;
-  opacity: 0;
-  border-top-width: 1px;
-}
-
-.faq-text.show{
-  height: auto;
-  margin-top: 0.5rem;
-  padding-top: 0.5rem;
-  opacity: 1;
 }
 
 @media (min-width: 1536px) {
