+++ app/Resources/css/contact/faq.section.css
@@ -1,0 +1,12 @@
+.faq-text {
+  height: 0;
+  opacity: 0;
+  border-top-width: 1px;
+}
+
+.faq-text.show {
+  height: auto;
+  margin-top: 0.5rem;
+  padding-top: 0.5rem;
+  opacity: 1;
+}
