+++ app/Resources/css/contact/form.section.css
@@ -1,0 +1,31 @@
+.bg-contact {
+  background: linear-gradient(
+    135deg,
+    hsl(176, 43%, 45%) 0%,
+    hsl(176, 43%, 47%) 40%,
+    hsl(176, 43%, 53%) 60%,
+    hsl(176, 43%, 55%) 100%
+  );
+
+  color: #f5f5f5;
+  border-bottom: 8px solid transparent;
+  border-image: repeating-linear-gradient(
+    to right,
+    #fafafa 0 10px,
+    transparent 10px 20px
+  );
+  border-image-slice: 1;
+}
+
+@media (min-width: 768px) {
+  .bg-contact {
+    border-bottom: 0px;
+    border-right: 8px solid transparent;
+    border-image: repeating-linear-gradient(
+      to bottom,
+      #fafafa 0 10px,
+      transparent 10px 20px
+    );
+    border-image-slice: 1;
+  }
+}
