+++ templates/contact/contact.form.html.twig
@@ -1,0 +1,70 @@
+<div class="max-w-3xl w-full border border-8 shadow overflow-hidden flex flex-col md:flex-row">
+
+    <div class="md:max-w-[17rem] lg:max-w-xs w-full px-4 py-8 sm:px-8 flex flex-col gap-6 bg-contact ">
+        <h1 class="text-3xl font-semibold">Get in Touch</h1>
+
+        <div>
+            <h2 class="text-lg font-semibold">Direct Chat</h2>
+            <p class="text-sm">Get a response within 24 hours.</p>
+            <p class="text-sm font-semibold">dev.jymarkb@gmail.com</p>
+        </div>
+
+        <div>
+            <h2 class="text-lg font-semibold">Direct Call</h2>
+            <p class="text-sm">Available Mon-Fri, from 10 AM to 8 PM.</p>
+            <p class="text-sm font-semibold">(+63) 962-2633-264</p>
+        </div>
+
+        <div>
+            <h3 class="text-lg font-semibold">Social Media</h3>
+            <ul class="flex gap-4 text-xl">
+                <li>
+                    <a href="{{ path('facebook') }}" target="_blank" rel="noopener noreferrer" aria-label="Open Facebook">
+                        <i class="icon-facebook"></i>
+                    </a>
+                </li>
+                <li>
+                    <a href="{{ path('github') }}" target="_blank" rel="noopener noreferrer" aria-label="Open GitHub">
+                        <i class="icon-github"></i>
+                    </a>
+                </li>
+                <li>
+                    <a href="{{ path('email') }}" aria-label="Open Email">
+                        <i class="icon-mail"></i>
+                    </a>
+                </li>
+                <li>
+                    <a href="{{ path('linkedin') }}" target="_blank" rel="noopener noreferrer" aria-label="Open LinkedIn">
+                        <i class="icon-linkedin"></i>
+                    </a>
+                </li>
+            </ul>
+        </div>
+    </div>
+
+    <div class="w-full px-4 py-8 sm:px-8 bg-white">
+
+        <form id="google-form" action="https://docs.google.com/forms/u/0/d/e/1FAIpQLScW-phm4J-cQKc2t65xEPLXGboE2y47zZaZQ1RT6j7a6zOkQw/formResponse" method="POST" class="flex flex-col gap-4">
+
+            <div class="flex flex-col sm:flex-row gap-4">
+                <input type="text" class="border border-input bg-white/50 px-3 text-base rounded-md shadow placeholder:text-muted-foreground text-sm h-10 w-full" placeholder="First Name" required name="entry.1583028837">
+                <input type="text" class="border border-input bg-white/50 px-3 text-base rounded-md shadow placeholder:text-muted-foreground text-sm h-10 w-full" placeholder="Last Name" required name="entry.454150918">
+            </div>
+
+            <input type="email" class="border border-input bg-white/50 px-3 text-base rounded-md shadow placeholder:text-muted-foreground text-sm h-10 w-full" placeholder="Email@domain.com" required name="entry.1006288430">
+
+            <select type="select" class="border border-input bg-white px-3 text-base rounded-md shadow text-sm h-10 w-full text-muted-foreground " required name="entry.482077460" aria-label="category">
+                <option class="hidden" value="" disabled selected="selected">Subject</option>
+                <option value="Content creation">Content creation</option>
+                <option value="Account">Account Issue</option>
+                <option value="Other">Other</option>
+            </select>
+
+            <textarea class="border border-input bg-white/50 px-3 text-base rounded-md shadow placeholder:text-muted-foreground text-sm" placeholder="Message . . ." name="entry.479743662" rows="8"></textarea>
+
+            <button class="flex items-center justify-center gap-2 bg-primaryTheme text-white text-sm py-2 rounded-md shadow" type="submit">Submit <i class="icon-send"></i></button>
+
+        </form>
+    </div>
+
+</div>
