+++ templates/contact/index.html.twig
@@ -15,30 +15,6 @@
 {% block title %}Contact | The &lt;/blog&gt; Blog
 {% endblock %}
 
-{% set  faqs = [
-	{
-		'question': 'What is this CMS built with?',
-		'answer': 'This CMS is built using Symfony 7 for the backend and React + TypeScript on the frontend. It uses Tailwind CSS for styling, TinyMCE for the editor, and Docker for containerization.'
-	},
-	{
-		'question': 'Can I customize the design and layout?',
-		'answer': 'Yes! The entire UI is built with Tailwind CSS and React, making it super easy to customize components, themes, or layout structure.'
-	},
-	{
-		'question': 'Is there a rich text editor?',
-		'answer': 'Definitely! We use TinyMCE for a seamless WYSIWYG writing experience. It supports links, images, formatting, and custom plugins.'
-	},
-	{
-		'question': 'Can I manage images and uploads?',
-		'answer': 'Yes, image uploading is supported.'
-	},
-	{
-		'question': 'Is this open source?',
-		'answer': 'Yes! You can view the source code, fork it, or contribute on <a href="' ~ path('github') ~ '" class="text-indigo-600 underline"  target="_blank" rel="noopener noreferrer" >GitHub</a>.'
-	}
-] %}
-
-
 {% block body %}
 
 	<main class="border-grid border-b">
@@ -55,110 +31,19 @@
 				
 				</section>
 
-				<div class="flex-1 h-2 bg-gray-800">
-				</div>
+				<div class="flex-1 h-2 bg-gray-800"></div>
 
 				<section class="absolute w-full top-40 md:top-44 lg:top-48 py-16 px-4 sm:px-12 lg:px-16 flex justify-center 2xl:w-fit 2xl:right-0">
-					<div class="max-w-3xl w-full border border-8 shadow overflow-hidden flex flex-col md:flex-row">
-
-						<div class="md:max-w-[17rem] lg:max-w-xs w-full px-4 py-8 sm:px-8 flex flex-col gap-6 bg-contact ">
-							<h1 class="text-3xl font-semibold">Get in Touch</h1>
-
-							<div>
-								<h2 class="text-lg font-semibold">Direct Chat</h2>
-								<p class="text-sm">Get a response within 24 hours.</p>
-								<p class="text-sm font-semibold">dev.jymarkb@gmail.com</p>
-							</div>
-
-							<div>
-								<h2 class="text-lg font-semibold">Direct Call</h2>
-								<p class="text-sm">Available Mon-Fri, from 10 AM to 8 PM.</p>
-								<p class="text-sm font-semibold">(+63) 962-2633-264</p>
-							</div>
-
-							<div>
-								<h3 class="text-lg font-semibold">Social Media</h3>
-								<ul class="flex gap-4 text-xl">
-									<li>
-										<a href="{{ path('facebook') }}" target="_blank" rel="noopener noreferrer" aria-label="Open Facebook">
-											<i class="icon-facebook"></i>
-										</a>
-									</li>
-									<li>
-										<a href="{{ path('github') }}" target="_blank" rel="noopener noreferrer" aria-label="Open GitHub">
-											<i class="icon-github"></i>
-										</a>
-									</li>
-									<li>
-										<a href="{{ path('email') }}" aria-label="Open Email">
-											<i class="icon-mail"></i>
-										</a>
-									</li>
-									<li>
-										<a href="{{ path('linkedin') }}" target="_blank" rel="noopener noreferrer" aria-label="Open LinkedIn">
-											<i class="icon-linkedin"></i>
-										</a>
-									</li>
-								</ul>
-							</div>
-						</div>
-
-						<div class="w-full px-4 py-8 sm:px-8 bg-white">
-
-							<form id="google-form" action="https://docs.google.com/forms/u/0/d/e/1FAIpQLScW-phm4J-cQKc2t65xEPLXGboE2y47zZaZQ1RT6j7a6zOkQw/formResponse" method="POST" class="flex flex-col gap-4">
-
-								<div class="flex flex-col sm:flex-row gap-4">
-									<input type="text" class="border border-input bg-white/50 px-3 text-base rounded-md shadow placeholder:text-muted-foreground text-sm h-10 w-full" placeholder="First Name" required name="entry.1583028837">
-									<input type="text" class="border border-input bg-white/50 px-3 text-base rounded-md shadow placeholder:text-muted-foreground text-sm h-10 w-full" placeholder="Last Name" required name="entry.454150918">
-								</div>
-
-								<input type="email" class="border border-input bg-white/50 px-3 text-base rounded-md shadow placeholder:text-muted-foreground text-sm h-10 w-full" placeholder="Email@domain.com" required name="entry.1006288430">
-
-								<select type="select" class="border border-input bg-white px-3 text-base rounded-md shadow text-sm h-10 w-full text-muted-foreground " required name="entry.482077460" aria-label="category">
-									<option class="hidden" value="" disabled selected="selected">Subject</option>
-									<option value="Content creation">Content creation</option>
-									<option value="Account">Account Issue</option>
-									<option value="Other">Other</option>
-								</select>
-
-								<textarea class="border border-input bg-white/50 px-3 text-base rounded-md shadow placeholder:text-muted-foreground text-sm" placeholder="Message . . ." name="entry.479743662" rows="8"></textarea>
-
-								<button class="flex items-center justify-center gap-2 bg-primaryTheme text-white text-sm py-2 rounded-md shadow" type="submit">Submit <i class="icon-send"></i></button>
-
-							</form>
-						</div>
-
-					</div>
+					{% include 'contact/contact.form.html.twig'%}
 				</section>
 
 				<section class="bg-gray-100 p-16 px-4 sm:px-8 md:px-12 lg:px-16 pt-[calc(1040px-17rem)] sm:pt-[calc(980px-16rem)] md:pt-[calc(606px-16rem)] lg:pt-[calc(606px-14rem)] 2xl:pt-16">
-
-					<div class="faq-container">
-						<h1 class="font-bold text-3xl text-gray-800 text-center 2xl:text-left">Frequently Asked Questions (FAQs)</h1>
-
-						<div id="faq-row-wrapper" class="flex flex-col gap-2 pt-4">
-
-						{% for faq in faqs %}
-
-						<div class="faq-row-container bg-white/50 shadow p-4 rounded ">
-							<div class="flex justify-between items-center">
-								<h2 class="font-semibold text-lg">{{ faq.question }}</h2>
-								<button class="transition duration-300 hover:bg-primaryTheme hover:text-white rounded" aria-label="dropdown button">
-									<i class="flex items-center justify-center h-8 w-8 font-bold icon-chevron-down"></i>
-								</button>
-							</div>
-
-							<p class=" transition-all duration-300 ease-in-out overflow-hidden faq-text {{ loop.first ? 'show' : '' }}">{{ faq.answer|raw }}</p>
-						</div>
-							
-						{% endfor %}
-						</div>
-					</div>
+	
+					{% include 'contact/contact.faq.html.twig'%}
 
 				</section>
+
 			</div>
-
-
 		
 		</div>
 	</main>
